[issue95] [Refactor] : social login code cleanup
