+++ src/main/kotlin/com/tinuproject/tinu/DTO/ResponseDTO.kt
@@ -5,5 +5,5 @@ import com.fasterxml.jackson.annotation.JsonProperty
 data class ResponseDTO(
         @JsonProperty("success") val isSuccess: Boolean,
         @JsonProperty("state") val stateCode: Int,
-        @JsonProperty("result") val result: Any
+        @JsonProperty("result") val result: Any?
 )
