+++ src/main/kotlin/com/tinuproject/tinu/GlobalExceptionHandler.kt
@@ -1,10 +1,7 @@
 package com.tinuproject.tinu
 
-import com.fasterxml.jackson.databind.ser.Serializers.Base
+import com.tinuproject.tinu.DTO.ResponseDTO
 import com.tinuproject.tinu.domain.exception.base.BaseException
-import com.tinuproject.tinu.domain.exception.base.ResponseDTO
-import com.tinuproject.tinu.domain.exception.token.InvalidedTokenException
-import com.tinuproject.tinu.domain.exception.token.NotFoundTokenException
 import com.tinuproject.tinu.web.ResponseEntityGenerator
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ExceptionHandler
