+++ src/main/kotlin/com/tinuproject/tinu/domain/exception/base/BaseErrorCode.kt
@@ -1,5 +1,7 @@
 package com.tinuproject.tinu.domain.exception.base
 
+import com.tinuproject.tinu.DTO.ResponseDTO
+
 interface BaseErrorCode {
     fun getResponse(): ResponseDTO?
 }
