--- src/main/kotlin/com/tinuproject/tinu/domain/exception/base/ErrorResponseDTO.kt
@@ -1,10 +1,0 @@
-package com.tinuproject.tinu.domain.exception.base
-
-class ErrorResponseDTO(
-    val isSuccess : Boolean,
-    val httpStatusCode : Int,
-    val errorCode : String,
-    val result : Any
-) {
-
-}
