+++ src/main/kotlin/com/tinuproject/tinu/domain/token/refreshtoken/controller/RefreshTokenController.kt
@@ -1,7 +1,6 @@
 package com.tinuproject.tinu.domain.token.refreshtoken.controller
 
-import com.tinuproject.tinu.domain.exception.base.ResponseDTO
-import com.tinuproject.tinu.domain.exception.token.InvalidedTokenException
+import com.tinuproject.tinu.DTO.ResponseDTO
 import com.tinuproject.tinu.domain.exception.token.NotFoundTokenException
 import com.tinuproject.tinu.domain.token.Tokens
 import com.tinuproject.tinu.domain.token.refreshtoken.service.RefreshTokenService
@@ -12,7 +11,6 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.CookieValue
@@ -35,29 +33,18 @@ class RefreshTokenController(
 
     @GetMapping("/generate")
     fun generateAccessToken(httpServletResponse: HttpServletResponse, @CookieValue(name = "RefreshToken") refreshToken : String?): ResponseEntity<ResponseDTO> {
-        var tokens : Tokens
         log.info("AccessToken 갱신 시도")
         if(refreshToken==null){
             throw NotFoundTokenException()
         }
 
-        tokens = refreshTokenService.reissueAccessTokenByRefreshToken(refreshToken)
+        val tokens : Tokens = refreshTokenService.reissueAccessTokenByRefreshToken(refreshToken)
 
 
         httpServletResponse.addHeader(HttpHeaders.SET_COOKIE,CookieGenerator.createCookies(accessTokenKey,tokens.accessToken))
         httpServletResponse.addHeader(HttpHeaders.SET_COOKIE,CookieGenerator.createCookies(refreshTokenkey, tokens.refreshToken))
         var body : MutableMap<String, Any> = mutableMapOf()
 
-        //TODO 이후 삭제 예정 잘 보내지는 지 responseBody를 통해 확인하기 위함.
-        body.put("Tokens",tokens)
-
-//        val result = ResponseDTO(
-//            isSuccess = true,
-//            httpStatusCode = HttpStatus.OK.ordinal,
-//            result = body
-//        )
-
-//        var responseEntity : ResponseEntity<ResponseDTO> = ResponseEntity.ok().body(result)
 
         val responseEntity = ResponseEntityGenerator.onSuccess(body)
         return responseEntity
